+++ priv/repo/migrations/20220901203547_create_increments.exs
@@ -4,7 +4,7 @@ defmodule Incrementer.Repo.Migrations.CreateIncrements do
   def change do
     create table(:increments) do
       add :key, :string, primary_key: true
-      add :value, :float
+      add :value, :integer
     end
 
     create unique_index(:increments, [:key])
